extract summarizer to util add abort handlers and controller for langchain stuff so socket close kills process
